Re-check surviving account is open before executing merge
Fold an A-open EXISTS guard into execute_merge's proposal-consume UPDATE
so a soft-closed surviving account makes the consume affect zero rows and
hit the existing Ok(None) no-commit path. The proposal FK only guarantees
A exists, not that it is open, and closed_at does not trigger ON DELETE
CASCADE, so without this guard B's principals and authenticators could be
folded onto a tombstoned account. Add execute_rejects_when_surviving_closed
asserting deny + no mutation. Widen the unshipped V34
absorbed_principal_count column to BIGINT and drop the i32 truncating
cast so struct, column, and count are all i64.

+++ migrations/V34__account_consolidation.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS trace_account_merge_proposals (
     proposal_id              UUID NOT NULL,
     surviving_account_id     UUID NOT NULL,
     absorbed_account_id      UUID NOT NULL,
-    absorbed_principal_count INT NOT NULL DEFAULT 0,
+    absorbed_principal_count BIGINT NOT NULL DEFAULT 0,
     created_at               TIMESTAMPTZ NOT NULL DEFAULT now(),
     expires_at               TIMESTAMPTZ NOT NULL,
     consumed_at              TIMESTAMPTZ,
